Experiments/requirements.txt
olefile==0.44
Pillow==4.2.1
pytesseract==0.1.7
